Log platform version
