+++ lib/shared/src/telemetry/EventLogger.ts
@@ -7,6 +7,8 @@ import { TelemetryEventProperties } from '.'
 export interface ExtensionDetails {
     ide: 'VSCode' | 'JetBrains' | 'Neovim' | 'Emacs'
     ideExtensionType: 'Cody' | 'CodeSearch'
+    platform: string
+    arch?: string
 
     /** Version number for the extension. */
     version: string
